// thirdeye-ui/src/app/components/alert-view/alert-drawer/alert-drawer.interfaces.ts
/*
 * Copyright 2024 StarTree Inc
 *
 * Licensed under the StarTree Community License (the "License"); you may not use
 * this file except in compliance with the License. You may obtain a copy of the
 * License at http://www.startree.ai/legal/startree-community-license
 *
 * Unless required by applicable law or agreed to in writing, software distributed under the
 * License is distributed on an "AS IS" BASIS, WITHOUT * WARRANTIES OF ANY KIND,
 * either express or implied.
 *
 * See the License for the specific language governing permissions and limitations under
 * the License.
 */
import { Alert } from "../../../rest/dto/alert.interfaces";

export interface AlertOption {
    id: string;
    label: string;
    onClick?: () => void;
    subOptions?: AlertOption[];
    component?: React.ReactNode;
}

export interface AlertOptionGroup {
    id: string;
    title?: string;
    options: AlertOption[];
}

export interface AlertDrawerProps {
    alert: Alert;
    onChange: (alert: Alert) => void;
    onDetectionRerunSuccess: () => void;
}